config: Don't require packages to be alphabetical
The configuration parser requries that
entries in the 'packages' section
are in alphabetical order.

It will fail parsing if that's not the case,
even if the configuration is otherwise valid.

This seems like an unnecessary artificial limitation.
Enforcing such a convention should be the user's choice.

This change deletes this limitation.