template: Use a more fluid layout
Instead of using a table, take advantage of the grid layout.
We still print a table of sorts,
but it's more fluid in appearance based on width of the screen.

On narrower screens, we'll show a listing
where each item has a description label next to it
rather than at the top.