ci: Publish Docker image to ghcr.io on release
Adds a workflow that builds and publishes a Docker image
to ghcr.io based on the Dockerfile in the root of the repository.

This will become available at
https://github.com/uber-go/sally/pkgs/container/sally.
Users will be able to import it by using `ghcr.io/uber-go/sally:$tag`
with `docker pull` or in the `FROM` clause of their own Docker file.

As a test for this, I published it manually to my own fork.
The image is at https://github.com/abhinav/sally/pkgs/container/sally.
With that, I ran the following Dockerfile:

    FROM ghcr.io/abhinav/sally:master
    COPY sally.yaml /

And it worked as expected.