refactor(indexHandler): Don't retain configuration
Similar to the previous patch,
this patch removes knowledge of the configuration from the index handler.
Instead, the index handler is now passed a list of packageInfo structs
each containing just the information needed to render the index page.

This also obviates the need to duplicate the logic to compute
the package import path in the index.html template.