Update fossa action to v1.1.0
The FOSSA check has been failing for a while with
fossa-actions@v1 Error: Input required and not supplied: api-key

This has caused the dependabot queue to backup, can we can't merge them
manually in sally because this check is required.

Updating the version seems to cause the check to run successfully again.

Other repositories have run into the same issue
ref: https://github.com/getsentry/self-hosted/issues/1351