Add support for sub-indexes
<details>
<summary>Background</summary>

Sally renders two kinds of pages:

- packages: These are for packages defined in sally.yaml
  and any route under the package path.
- indexes: These list available packages.

The latter--indexes was previously only supported at '/', the root page.

This leads to a slight UX issue:
if you have a package with a / in its name (e.g. net/metrics):

- example.com/net/metrics gives you the package page
- example.com/ lists net/metrics
- However, example.com/net fails with a 404

</details>

This adds support for index pages on all parents of package pages.
Therefore, if example.com/net/metrics exists,
example.com/net will list all packages defined under that path.

To make this possible, the internals have been rewritten to some degree:
CreateHandler no longer builds an http.ServeMux
with a bunch of pre-registered http.Handlers.
It now builds a custom http.Handler with its own routing.

Routing is backed by a new pathTree abstraction.
pathTree is a specialized string-value storage
with support for inheriting values from a parent.
It has the following methods:

- `Set(path, v)` associates v with path.
- `Lookup(path)` returns the value at path,
  or the value inherited from the closest ancestor.
- `ListByPath(path)` lists all values at or under the given path.

For additional confidence in the implementation of pathTree,
a property test based on rapid is included.

Resolves #31